Working User Rating updates!

+++ cs564proj/DataLibrary/UserRating.cs
@@ -4,7 +4,7 @@ namespace DataLibrary
 {
     public class UserRating
     {
-        [Name("userID")] public int UserId { get; set; }
+        [Name("userID")] public long UserId { get; set; }
         [Name("movieID")] public int MovieId { get; set; }
         [Name("rating")] public double Rating { get; set; }
         [Name("timestamp")] public ulong Timestamp { get; set; }
